requirements.txt
# RAG Chatbot - Free Stack (No API keys needed)
# Install with: pip install -r requirements.txt

# Core LangChain packages
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
langchain-core==0.2.38

# Vector Database (local, free)
chromadb==0.5.5

# Document loaders
pypdf==4.3.1
unstructured==0.15.0
docx2txt==0.8

# Text processing
tiktoken==0.7.0

# Web UI
streamlit==1.38.0

# Evaluation (optional, Day 3)
ragas==0.1.21
datasets==2.21.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
